--- src/app/landing/layout.tsx
@@ -1,27 +1,0 @@
-import "~/styles/globals.css";
-
-import { GeistSans } from "geist/font/sans";
-import { type Metadata } from "next";
-
-import Footer from "../_components/layout/Footer";
-import { Navbar } from "../_components/layout/navbar";
-
-export const metadata: Metadata = {
-  title: "OpenTask",
-  description: "Micro-task Management on the Blockchain",
-  icons: [{ rel: "icon", url: "/logo.svg" }],
-};
-
-export default function RootLayout({
-  children,
-}: Readonly<{ children: React.ReactNode }>) {
-  return (
-    <html lang="en" className={`${GeistSans.variable}`}>
-      <body>
-        <Navbar />
-        {children}
-        <Footer />
-      </body>
-    </html>
-  );
-}
