+++ src/hooks/sidebarAtom.ts
@@ -1,4 +1,4 @@
-// src/atoms/sidebarAtom.ts
+
 import { atom } from "jotai";
 
-export const sidebarAtom = atom(false); // false = closed, true = open
+export const sidebarAtom = atom(true);
